Use a simpler regexp for the email validation example
This one is intended to showcase how to use the format validation, and
also as a simple way to check for possible email types, not necessarily
to be used as-is or to be an extensive email format validation regexp.

This regexp was copied over from Devise (while also allowing case
insensitive values, which Devise handles on its own). If you have the
need to cover more special cases, please use some sort of email
validator library or a different regexp.

Closes #70.