Fix Active Record integration
To integrate with Active Record you just need to
`include MailForm::Delivery` and implement the `headers` method. An
email should be sent whenever the model is created successfully.

That was broken since we introduced these extra conditionals to fix the
delivery of MailForm, due to the change in how Rails handles callbacks
abortion to not rely on the `false` value returning from `before_*`
callbacks, to explicitly require the code to `throw(:abort)` instead.

This was happening for objects inheriting from `MailForm::Base` so it
was working fine there, but it broke Active Record integrations since it
was just doing that for the `*_deliver` callbacks, and because of the
way the check was introduced, it was trying to use those same callbacks
inside AR, and they don't exist there, resulting in errors:

    NoMethodError (undefined method `before_deliver'...

This change to callbacks was introduced by Rails 5
(https://github.com/rails/rails/pull/17227), which means we can safely
change the logic to `throw(:abort)` for both `deliver` and `create`
callbacks going forward.

Include some tests that simulate Active Record behavior through Active
Model and `create` callbacks, since we didn't have any specific test
coverage there.

Closes #56.